Upload bug fix. Added functionality to report server errors. Graph bug fix.

+++ web/POC/scripts/public_api.php
@@ -12,46 +12,10 @@ if (isset($_GET["upload"])) {
 		if ($rna)
 			$rna->json();
 		else
-			echo "Error: parsing file.";
-	} else{
-		echo "There was an error uploading the file, please try again!";
-	}
+			echo json_encode(array("errors"=>"Error: parsing file: ".$_FILES[$field_name]['name'].", please try again!"));
+	} else 
+		echo json_encode(array("errors"=>"Server error uploading file ".$_FILES[$field_name]['name'].", please try again!"));
 }
-
-/*
-if (isset($_GET["zip"])) {
-	$stamp = time();
-	$dir = "../uploads/".$stamp."/";
-	mkdir($dir);
-	$filename = basename( $_FILES[$field_name]['name']);
-	$target_path = $dir . $filename;
-	if(move_uploaded_file($_FILES[$field_name]['tmp_name'], $target_path)) {
-		if (extract_zip($target_path, $dir)) {
-			unlink($target_path);
-			$root = scandir($dir);
-			$rna_collection = array();
-			foreach($root as $value)
-			{
-				if($value === '.' || $value === '..')
-					continue;
-				if(is_file("$dir/$value")) {
-					$rna = parse_ct_file($dir . $value, $value);
-					if ($rna)
-						$rna_collection[] = $rna->as_array();
-					else 
-						echo "<br />Error: parsing file - '" . $dir . $value . "' - '" . $rna . "'";
-				}
-				unlink("$dir/$value");
-			}
-			rmdir($dir);
-			echo json_encode($rna_collection);
-		}
-	} else{
-		echo "There was an error uploading the file, please try again!";
-	}
-}
-/**/
-
 
 if (isset($_GET["zip"])) {
 	$stamp = time();
@@ -64,7 +28,7 @@ if (isset($_GET["zip"])) {
 			unlink($target_path);
 			// first level of zip
 			$ts_collection = array();
-			$base = parse_training_set($dir, $filename);
+			$base = parse_training_set($dir, $filename, 1);
 			if ($base)
 				$ts_collection[] = $base;
 			
@@ -75,7 +39,7 @@ if (isset($_GET["zip"])) {
 				if($value === '.' || $value === '..')
 					continue;
 				if(is_dir("$dir/$value")) {
-					$ts = parse_training_set("$dir/$value", $value);
+					$ts = parse_training_set("$dir/$value", $value, 0);
 					if ($ts) {
 						$ts_collection[] = $ts;
 					}
@@ -88,32 +52,45 @@ if (isset($_GET["zip"])) {
 			echo json_encode($ts_collection);
 		}
 	} else{
-		echo "There was an error uploading the file, please try again!";
+		echo json_encode(array("errors"=>"Server error uploading file ".$_FILES[$field_name]['name'].", please try again!"));
 	}
+	
 }
 /**/
 
-function parse_training_set($dir, $name) {
+function parse_training_set($dir, $name, $is_base) {
 	$root = scandir($dir);
 	$ts = new TrainingSet($name, 0, 0);
 	foreach($root as $value)
 	{
 		if($value === '.' || $value === '..')
 			continue;
-		if(is_file("$dir/$value")) {
-			$ts->add_seq(parse_ct_file("$dir/$value", $value));
+		if (is_dir("$dir/$value") && !$is_base) {
+			// Invalid directory structure
+			// TODO: Return error message
+			break;
+		}
+		if(is_file("$dir/$value") && strpos($value, "ct") == (strlen($value) - 2)) {
+			$seq = parse_ct_file("$dir/$value", $value);
+			if ($seq)
+				$ts->add_seq($seq);
+			else
+				$error_list[] = $error;
 			unlink("$dir/$value");
 		}
 		// TODO: validate input!!
 		if($value == 'info.txt') {
-			$file_handle = fopen($value, "r");
+			$file_handle = fopen("$dir/$value", "r");
 			if (!feof($file_handle)) {
 				$ts->name = fgets($file_handle);
 				if (!feof($file_handle))
 					$ts->color = fgets($file_handle);
 			}
 			fclose($file_handle);
+			unlink("$dir/$value");
 		}
+		if (file_exists("$dir/$value") && is_file("$dir/$value"))
+			unlink("$dir/$value");
 		/**/
 	}
 	//rmdir($dir);
